requirements.txt
transliterate==1.10.2
num2words==0.5.10
Pillow==10.0.0